refactor(common): putting values in their own files
