+++ icann-rdap-common/src/response/values/entity_role.rs
@@ -1,0 +1,20 @@
+//! IANA registered roles for entities.
+
+use super::EnumDisplay;
+use strum::EnumString;
+
+#[derive(PartialEq, Eq, Debug, EnumString, EnumDisplay)]
+#[strum(serialize_all = "lowercase")]
+pub enum EntityRole {
+    Registrant,
+    Technical,
+    Administrative,
+    Abuse,
+    Billing,
+    Registrar,
+    Reseller,
+    Sponsor,
+    Proxy,
+    Notifications,
+    Noc,
+}
